nomad agent refactor
